feat(wasm-visor): serve /api/network-view in the wasm core (was 404)
The wasm core's /api router didn't implement network-view (the native HV serves
it), so the network-view table + the visualizer rendered empty in the browser.

- Extract the SD/TPD/UT aggregation into a dependency-free leaf pkg/visor/netview
  (Compute(fetch)) — pkg/visor doesn't compile for js/wasm, so the wasm tab can't
  import it; the leaf only needs encoding/json+strings+time. pkg/visor aliases the
  types (NetworkViewEntry=netview.Entry, NetworkViewResponse=netview.Response) and
  delegates, so native + browser visors share ONE aggregation and can't drift.
- SelfProvider gains SelfNetworkView() []byte; visorSelf builds it from the tab's
  own dmsg fetch (FetchOverDmsg to sd/tpd/ut, reachable since the service-PK seed
  fix) via netview.Compute. router.go serves /network-view from it.

Verified via hvinspect: the wasm network-view table now renders real entries
(row-offline/row-not-in-ut) where it was empty.
