+++ pkg/wasmhv/self_test.go
@@ -25,6 +25,7 @@ func (f fakeSelf) SelfSummary() Summary {
 }
 func (f fakeSelf) SelfTransports() []*TransportSummary { return f.tps }
 func (f fakeSelf) SelfRoutes() []byte                  { return []byte("[]") }
+func (f fakeSelf) SelfNetworkView() []byte             { return nil }
 
 func newSelfPK(t *testing.T) cipher.PubKey {
 	t.Helper()
